webui/requirements.txt
flask>=2.0.0
gunicorn>=20.0.0
boto3>=1.26.0
google-cloud-pubsub>=2.13.0
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.7.2
botocore==1.34.14
Werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
google-auth==2.23.4
setuptools==68.2.2
requests==2.31.0
python-dotenv>=0.19.0
pyyaml>=6.0 